refactor(evolution): split out evaluator_manifest module (R6)
Per §5 C1 of the scope report (helix-evolution-refactor-scope.md):
the 7 manifest functions + 7 protected-file copy helpers (~313 LOC
including associated constants) had zero dependency on EvolutionState,
ParetoFrontier, or the evolution loop — they operate only on Path,
HelixConfig, and Candidate.worktree_path.  Extracted to a new module
that can be unit-tested in isolation without constructing evolution-loop
state.

New module: src/helix/evaluator_manifest.py (374 LOC)

14 symbols moved:
  Protected-file helpers: _extract_script_token, _looks_like_script_file,
    _to_repo_relative, _collect_protected_evaluator_paths,
    _copy_protected_path, _refresh_protected_evaluator_files,
    _refresh_and_snapshot_protected_evaluator_files
  Manifest helpers: _sha256_file, _iter_protected_manifest_files,
    _evaluator_manifest_path, _build_evaluator_integrity_manifest,
    _write_evaluator_integrity_manifest,
    _load_evaluator_integrity_manifest, _detect_evaluator_tamper

Backward compatibility: evolution.py re-exports all 14 symbols via
  `from helix.evaluator_manifest import ...`
so any code importing them from `helix.evolution` continues to work.

Associated constants (_PROTECTED_DIRECTORY_IGNORE_PATTERNS,
_PROTECTED_DIRECTORY_IGNORE_MATCHER, _NO_SCRIPT_COMMANDS, etc.) moved
with the functions they exclusively support.  The three constants still
needed by _check_evaluator_script_exists (which stays in evolution.py)
are included in the re-import list.

Test import update:
  tests/unit/test_evaluator_integrity.py: `from helix.evolution import`
  → `from helix.evaluator_manifest import` (8 symbols)

Net: −313 LOC from evolution.py, +374 LOC new module (~+61 net to repo
from the module docstring).  Test suite: 815 passed.
